requirements.txt
aiogram
python-dotenv
requests
numpy
gino